Move execution-model-independent definitions up
