+++ src/i2c/blocking.rs
@@ -99,22 +99,7 @@
 //! }
 //! ```
 
-use crate::private;
-
-/// Address mode (7-bit / 10-bit)
-///
-/// Note: This trait is sealed and should not be implemented outside of this crate.
-pub trait AddressMode: private::Sealed {}
-
-/// 7-bit address mode type
-pub type SevenBitAddress = u8;
-
-/// 10-bit address mode type
-pub type TenBitAddress = u16;
-
-impl AddressMode for SevenBitAddress {}
-
-impl AddressMode for TenBitAddress {}
+pub use super::{AddressMode, SevenBitAddress, TenBitAddress};
 
 /// Blocking read
 pub trait Read<A: AddressMode = SevenBitAddress> {
