+++ src/timer/mod.rs
@@ -1,3 +1,6 @@
 //! Timer traits
 
 pub mod nb;
+
+/// Marker trait that indicates that a timer is periodic
+pub trait Periodic {}
